refactor(topology)!: move CLI dispatch into the standard crate as TopologyCommandHandler (#68)
The seven topology commands (analyze, validate, diff, check, comment, report, viz) now live in code_topology::cli behind the CL01 CommandHandler contract; register() exposes them to composed consumer binaries. aps-cli delegates through the same handler (main.rs 3581 to 1183 LOC, advances #36). vsa_config moves into the standard with its 16 tests. Verbose crosses the trait boundary via APSS_VERBOSE. viz degrades to exit 5 naming the missing feature when viz features are off.
